GH-6 format, organize packages.
I moved the meta related code outside of the core package. I made a
`_test` package that only contains tests. I also renamed a bunch of the
public items in the new meta package, because this package will be more
visible than other packages (plugins will have to import this package).
I also made some things private that didn't need to be public.
